fix(train): epoch accumulators to int64 (int32 overflow zeroed train_loss)
_epoch_valid_jax / _epoch_correct_jax started as jnp.int32. At 76k
steps × ~65k valid tokens / step the running sum crosses int32's 2.15e9
ceiling and wraps to negative; the `if epoch_valid > 0` guard at
end-of-epoch then falls through to the 0.0 fallback, which is why the
epoch summary printed "Train loss=0.0000, Train acc=0.0000" for an
otherwise healthy run.

Window accumulators stay int32 — they reset every LOG_INTERVAL=100
steps so at most ~6.5M tokens accumulate, nowhere near overflow.

+++ scripts/train_jax.py
@@ -2547,8 +2547,10 @@ def main():
         # Epoch accumulators on device — one device_get per epoch at the
         # end, rather than per-step float()/int() sync.
         _epoch_loss_jax = jnp.float32(0.0)
-        _epoch_correct_jax = jnp.int32(0)
-        _epoch_valid_jax = jnp.int32(0)
+        # int64: valid_count sums to ~n_steps * tokens_per_step, which
+        # exceeds int32 range (2.15e9) on any multi-billion-token epoch.
+        _epoch_correct_jax = jnp.int64(0)
+        _epoch_valid_jax = jnp.int64(0)
         epoch_steps = 0
 
         # Window accumulators on device — one device_get per log boundary.
